added user sidebar, renamed previous user bloc and refactored user bloc to match feed_bloc and community_bloc
