+++ lib/community/utils/post_card_action_helpers.dart
@@ -25,6 +25,7 @@ import 'package:thunder/shared/picker_item.dart';
 import 'package:thunder/shared/snackbar.dart';
 import 'package:thunder/thunder/bloc/thunder_bloc.dart';
 import 'package:thunder/user/bloc/user_bloc.dart';
+import 'package:thunder/user/enums/user_action.dart';
 import 'package:thunder/utils/instance.dart';
 import 'package:thunder/instance/utils/navigate_instance.dart';
 import 'package:thunder/user/utils/navigate_user.dart';
@@ -417,7 +418,7 @@ void onSelected(BuildContext context, PostCardAction postCardAction, PostViewMed
                 );
       break;
     case PostCardAction.blockUser:
-      context.read<UserBloc>().add(BlockUserEvent(personId: postViewMedia.postView.creator.id, blocked: true));
+      context.read<UserBloc>().add(UserActionEvent(userAction: UserAction.block, userId: postViewMedia.postView.creator.id, value: true));
       break;
     case PostCardAction.subscribeToCommunity:
       context.read<CommunityBloc>().add(CommunityActionEvent(communityAction: CommunityAction.follow, communityId: postViewMedia.postView.community.id, value: true));
